+++ src/docs/views/components/content/page-none-found/fragments/playground-none-found.vue
@@ -4,10 +4,6 @@
 			None found
 		</template>
 
-		<template #introduction>
-			<p />
-		</template>
-
 		<none-found>
 			<template #title>
 				{{ slots.title?.value }}